feat: added new library pavao-sys
