+++ pavao/src/smb/types/file.rs
@@ -5,11 +5,11 @@
 use std::io::{self, Read, Seek, SeekFrom, Write};
 
 use libc::{c_int, c_void, mode_t, off_t};
-
-use crate::libsmbclient::{
+use pavao_sys::{
     smbc_getFunctionClose, smbc_getFunctionLseek, smbc_getFunctionRead, smbc_getFunctionWrite,
     SMBCFILE,
 };
+
 use crate::{utils, SmbClient};
 
 pub struct SmbFile<'a> {
